Created new roles for ota-ce on the demo server
- install-prereqs
This role should be run just after the server has been wiped, but can be run at any
subsequent time. It makes sure that the necessary packages for running ota-ce are installed
on the demo server.

- configure-demo-server
This role should only be run after the server has been wiped and all the necessary
packages are installed on the demo-server. It will throw error if you run this role
after this role has been already run once. It configures the ota-ce to  be capable
of starting all its services on the demo-server and setup the landing page for the server.

- start-server
This role should be run after the demo server is configured and after a soft-wipe.
It will start all the services in docker containers and create key material, making
sure the services are online and available to use for the internet.

- soft-wipe-demo-server
This role should be run after the ota-ce services have been started.
This role will tear down all services and delete all key material and user data.