Created new playbooks
- hard-reset.yml
This playbook will reset the server completely and can be run anytime.
It starts by wiping out everyting on the server, then bringing up the firewall
and then installing all the necessary software to run ota-ce. Then it configures
ota-ce and runs the services.

- setup-server.yml
This playbook will wipe reset the server completely and then configure ota-ce on it.
It does exactly what hard-reset playbook does except it will not start the services
on the server. It will make sure that ota-ce is configured to run on the server.

- soft-reset.yml
This playbook will reset the server and start it again without wiping out the whole server.
It will bring down all the services and delete all key material, then it will start up those
services again.

- demo-uptane.yml
This playbook can be run once the server is completely wiped out and the firewall is up.
will install all the prerequisite software, configure the server to run ota-ce
and then start the services on the server. If all services are down and only the uptane
directory is deleted from the server, this playbook can be run to set uptane back up again.